refactor: densify units, production, trade, and train tests (#4642)
Concern-split Slice C suites so each file stays under 330 lines while
keeping existing panel harnesses as the source of truth.
