+++ app/test/features/game/widgets/civilian_units_sort_test.dart
@@ -26,37 +26,14 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app/features/game/widgets/units/civilian/civilian_units_sort.dart';
 
-const _humanId = 'gp1';
-const _otherId = 'gp2';
-
-Game _gameWith({
-  List<Unit> oldUnits = const [],
-  List<Unit> newUnits = const [],
-  List<Province> oldProvinces = const [],
-  List<Province> newProvinces = const [],
-}) {
-  return Game(
-    id: 'g',
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: RegionData(provinces: oldProvinces, units: oldUnits),
-      newWorld: RegionData(provinces: newProvinces, units: newUnits),
-    ),
-    players: const [
-      Player(id: _humanId, displayName: 'Human', isHuman: true),
-      Player(id: _otherId, displayName: 'Other', isHuman: false),
-    ],
-    minorNations: const [],
-    tribes: const [],
-  );
-}
+import 'civilian_units_sort_test_support.dart';
 
 void main() {
   suppressLogsForTests();
 
   group('provinceNamesByPrefixedId', () {
     test('indexes provinces from both regions and prefixes by regionId', () {
-      final game = _gameWith(
+      final game = civilianSortTestGameWith(
         oldProvinces: const [
           Province(
             id: 'oldWorld|p1',
@@ -82,23 +59,26 @@ void main() {
       expect(names, hasLength(3));
     });
 
-    test('iterates old-world entries before new-world (allProvinces order)', () {
-      final game = _gameWith(
-        oldProvinces: const [
-          Province(id: 'oldWorld|p1', regionId: 'oldWorld'),
-        ],
-        newProvinces: const [
-          Province(id: 'newWorld|p1', regionId: 'newWorld'),
-        ],
-      );
+    test(
+      'iterates old-world entries before new-world (allProvinces order)',
+      () {
+        final game = civilianSortTestGameWith(
+          oldProvinces: const [
+            Province(id: 'oldWorld|p1', regionId: 'oldWorld'),
+          ],
+          newProvinces: const [
+            Province(id: 'newWorld|p1', regionId: 'newWorld'),
+          ],
+        );
 
-      final keys = provinceNamesByPrefixedId(game).keys.toList();
+        final keys = provinceNamesByPrefixedId(game).keys.toList();
 
-      expect(keys, ['oldWorld|oldWorld|p1', 'newWorld|newWorld|p1']);
-    });
+        expect(keys, ['oldWorld|oldWorld|p1', 'newWorld|newWorld|p1']);
+      },
+    );
 
     test('returns an empty index when neither region has provinces', () {
-      final names = provinceNamesByPrefixedId(_gameWith());
+      final names = provinceNamesByPrefixedId(civilianSortTestGameWith());
 
       expect(names, isEmpty);
     });
@@ -106,11 +86,11 @@ void main() {
 
   group('isCivilianUnit', () {
     Unit civilianUnit(String id, String type) => Unit(
-          id: id,
-          type: type,
-          ownerId: _humanId,
-          locationProvinceId: 'oldWorld|p1',
-        );
+      id: id,
+      type: type,
+      ownerId: civilianSortHumanId,
+      locationProvinceId: 'oldWorld|p1',
+    );
 
     test('returns true for Builder, Explorer, Merchant, Engineer, Spy', () {
       expect(isCivilianUnit(civilianUnit('u1', kUnitTypeBuilder)), isTrue);
@@ -141,13 +121,13 @@ void main() {
         final unit = Unit(
           id: 'u1',
           type: kUnitTypeBuilder,
-          ownerId: _humanId,
+          ownerId: civilianSortHumanId,
           locationProvinceId: 'oldWorld|p9',
           tileKey: 'oldWorld|p9|0|0',
         );
         final name = civilianSortProvinceName(
           unit,
-          humanPlayerId: _humanId,
+          humanPlayerId: civilianSortHumanId,
           currentOrders: const Orders(),
           provinceNames: provinceNames,
         );
@@ -159,13 +139,13 @@ void main() {
       final unit = Unit(
         id: 'u1',
         type: kUnitTypeBuilder,
-        ownerId: _humanId,
+        ownerId: civilianSortHumanId,
         locationProvinceId: 'oldWorld|p1',
         tileKey: 'oldWorld|p1|0|0',
       );
       final name = civilianSortProvinceName(
         unit,
-        humanPlayerId: _humanId,
+        humanPlayerId: civilianSortHumanId,
         currentOrders: const Orders(),
         provinceNames: provinceNames,
       );
@@ -176,13 +156,13 @@ void main() {
       final unit = Unit(
         id: 'u1',
         type: kUnitTypeExplorer,
-        ownerId: _humanId,
+        ownerId: civilianSortHumanId,
         locationProvinceId: 'oldWorld|p1',
         tileKey: 'oldWorld|p1|0|0',
       );
       const orders = Orders(
         workOrdersByPlayerId: {
-          _humanId: [
+          civilianSortHumanId: [
             WorkOrder(
               unitId: 'u1',
               target: kWorkTargetExplore,
@@ -193,178 +173,11 @@ void main() {
       );
       final name = civilianSortProvinceName(
         unit,
-        humanPlayerId: _humanId,
+        humanPlayerId: civilianSortHumanId,
         currentOrders: orders,
         provinceNames: provinceNames,
       );
       expect(name, 'Cibola');
-    });
-  });
-
-  group('civilianUnitsInRegion', () {
-    const provinceNames = {
-      'oldWorld|oldWorld|p1': 'Avalon',
-      'oldWorld|oldWorld|p2': 'Belgrad',
-    };
-
-    Unit civilian({
-      required String id,
-      required String type,
-      required String tileKey,
-      String ownerId = _humanId,
-    }) {
-      return Unit(
-        id: id,
-        type: type,
-        ownerId: ownerId,
-        locationProvinceId: Unit.provinceIdFromTileKey(tileKey) ?? 'oldWorld|p1',
-        tileKey: tileKey,
-      );
-    }
-
-    test(
-      'orders results by province name, then type, then id (Schwartzian)',
-      () {
-        // Intentional input order is unsorted so the sort is observable.
-        final units = [
-          civilian(
-            id: 'u_zeta',
-            type: kUnitTypeExplorer,
-            tileKey: 'oldWorld|p2|0|0',
-          ),
-          civilian(
-            id: 'u_alpha',
-            type: kUnitTypeBuilder,
-            tileKey: 'oldWorld|p1|0|0',
-          ),
-          civilian(
-            id: 'u_beta',
-            type: kUnitTypeBuilder,
-            tileKey: 'oldWorld|p1|0|0',
-          ),
-          civilian(
-            id: 'u_gamma',
-            type: kUnitTypeExplorer,
-            tileKey: 'oldWorld|p1|1|0',
-          ),
-        ];
-
-        final sorted = civilianUnitsInRegion(
-          units,
-          _humanId,
-          provinceNames,
-          const Orders(),
-        );
-
-        expect(
-          sorted.map((u) => u.id).toList(),
-          ['u_alpha', 'u_beta', 'u_gamma', 'u_zeta'],
-        );
-      },
-    );
-
-    test(
-      'skips wrong owner, military, missing tile, and non-civilian units',
-      () {
-        final units = [
-          civilian(
-            id: 'foreign',
-            type: kUnitTypeBuilder,
-            tileKey: 'oldWorld|p1|0|0',
-            ownerId: _otherId,
-          ),
-          Unit(
-            id: 'no-tile',
-            type: kUnitTypeBuilder,
-            ownerId: _humanId,
-            locationProvinceId: 'oldWorld|p1',
-          ),
-          civilian(
-            id: 'military',
-            type: 'grenadiers',
-            tileKey: 'oldWorld|p1|0|0',
-          ),
-          civilian(
-            id: 'unknown-type',
-            type: 'aliens',
-            tileKey: 'oldWorld|p1|0|0',
-          ),
-          civilian(
-            id: 'civilian',
-            type: kUnitTypeBuilder,
-            tileKey: 'oldWorld|p1|0|0',
-          ),
-        ];
-
-        final sorted = civilianUnitsInRegion(
-          units,
-          _humanId,
-          provinceNames,
-          const Orders(),
-        );
-
-        expect(sorted.map((u) => u.id).toList(), ['civilian']);
-      },
-    );
-
-    test(
-      'sort key tracks projected tile when a pending order moves a unit',
-      () {
-        // u1 is currently in p1 ("Avalon"). A pending Explore order sends it
-        // to p2 ("Belgrad"), so it must sort *after* u2 (in p1) by province
-        // name even though `u1` is alphabetically earlier than `u2`.
-        final units = [
-          civilian(
-            id: 'u1',
-            type: kUnitTypeExplorer,
-            tileKey: 'oldWorld|p1|0|0',
-          ),
-          civilian(
-            id: 'u2',
-            type: kUnitTypeExplorer,
-            tileKey: 'oldWorld|p1|1|0',
-          ),
-        ];
-        const orders = Orders(
-          workOrdersByPlayerId: {
-            _humanId: [
-              WorkOrder(
-                unitId: 'u1',
-                target: kWorkTargetExplore,
-                targetTileKey: 'oldWorld|p2|0|0',
-              ),
-            ],
-          },
-        );
-
-        final sorted = civilianUnitsInRegion(
-          units,
-          _humanId,
-          provinceNames,
-          orders,
-        );
-
-        expect(sorted.map((u) => u.id).toList(), ['u2', 'u1']);
-      },
-    );
-
-    test('returns empty list when no units match', () {
-      final units = [
-        civilian(
-          id: 'military',
-          type: 'grenadiers',
-          tileKey: 'oldWorld|p1|0|0',
-        ),
-      ];
-
-      final sorted = civilianUnitsInRegion(
-        units,
-        _humanId,
-        provinceNames,
-        const Orders(),
-      );
-
-      expect(sorted, isEmpty);
     });
   });
 }
