+++ app/test/units_panel_shared_widgets_test.dart
@@ -283,82 +283,4 @@ void main() {
       expect(find.text('No civilian units'), findsNothing);
     });
   });
-
-  group('UnitsEntityActionRow', () {
-    testWidgets('renders details with text action label on wide width', (
-      WidgetTester tester,
-    ) async {
-      await pumpUnitsEntityActionRow(
-        tester,
-        actions: [unitsEntityMoveAction()],
-      );
-
-      expect(find.text('Left details'), findsOneWidget);
-      expect(find.text('Move'), findsOneWidget);
-      expect(find.byType(UnitsPanelRowChrome), findsOneWidget);
-    });
-
-    testWidgets('switches action button to icon-only on narrow width', (
-      WidgetTester tester,
-    ) async {
-      await pumpUnitsEntityActionRow(
-        tester,
-        width: 220,
-        actions: [unitsEntityMoveAction()],
-      );
-
-      expect(find.byIcon(Icons.route), findsOneWidget);
-      expect(find.text('Move'), findsNothing);
-    });
-
-    testWidgets(
-      'renders mockup compact-pill family per action variant (#3514): '
-      'neutral -> CtActionTextButton, danger -> CtDangerTextButton, '
-      'iconOnly -> CtCircularLocateButton, and no CtNinePatchButton',
-      (WidgetTester tester) async {
-        await pumpUnitsEntityActionRow(
-          tester,
-          actions: [
-            unitsEntityMoveAction(),
-            UnitsEntityAction(
-              tooltip: 'Cancel',
-              icon: Icons.cancel_outlined,
-              label: 'Cancel',
-              variant: UnitsEntityActionVariant.danger,
-              onPressed: () {},
-            ),
-            UnitsEntityAction(
-              tooltip: 'Locate',
-              icon: Icons.my_location,
-              label: 'Locate',
-              iconOnly: true,
-              onPressed: () {},
-            ),
-          ],
-        );
-
-        expect(find.byType(CtActionTextButton), findsOneWidget);
-        expect(find.byType(CtDangerTextButton), findsOneWidget);
-        expect(find.byType(CtCircularLocateButton), findsOneWidget);
-        expect(find.byType(CtNinePatchButton), findsNothing);
-        // The iconOnly Locate control renders no text label (circular pill).
-        expect(find.text('Locate'), findsNothing);
-      },
-    );
-
-    testWidgets(
-      'disabled action (onPressed == null) renders a disabled pill (#3514)',
-      (WidgetTester tester) async {
-        await pumpUnitsEntityActionRow(
-          tester,
-          actions: [unitsEntityMoveAction(enabled: false)],
-        );
-
-        final pill = tester.widget<CtActionTextButton>(
-          find.byType(CtActionTextButton),
-        );
-        expect(pill.onPressed, isNull);
-      },
-    );
-  });
 }
